+++ libs/ffdc-apis/corporate-accounts/CHANGELOG.md
@@ -1,1 +1,15 @@
 # Changelog
+
+<!-- TOC -->
+
+- [Changelog](#changelog)
+  - [0.1.1 (2020-05-19)](#011-2020-05-19)
+    - [Bug fixes](#bug-fixes)
+
+<!-- /TOC -->
+
+## 0.1.1 (2020-05-19)
+
+### Bug fixes
+
+- Fix interfaces not being exported correctly