+++ libs/oidc/CHANGELOG.md
@@ -4,24 +4,49 @@
 
 <!-- TOC depthFrom:2 depthTo:3 -->
 
-- [0.4.1 (2020-05-19)](#041-2020-05-19)
+- [0.5.0 (2020-05-19)](#050-2020-05-19)
+  - [Features](#features)
   - [Documentation](#documentation)
 - [0.4.0 (2020-05-06)](#040-2020-05-06)
-  - [Features](#features)
+  - [Features](#features-1)
 - [0.3.1 (2020-05-06)](#031-2020-05-06)
   - [Bug fixes](#bug-fixes)
 - [0.3.0 (2020-04-30)](#030-2020-04-30)
-  - [Features](#features-1)
+  - [Features](#features-2)
 - [0.2.1 (2020-04-30)](#021-2020-04-30)
   - [Bug fixes](#bug-fixes-1)
 - [0.2.0 (2020-04-23)](#020-2020-04-23)
-  - [Features](#features-2)
+  - [Features](#features-3)
 - [0.1.2 (2020-04-23)](#012-2020-04-23)
   - [Bug fixes](#bug-fixes-2)
 
 <!-- /TOC -->
 
-## 0.4.1 (2020-05-19)
+## 0.5.0 (2020-05-19)
+
+### Features
+
+- Module registration was harmonized with the rest of the modules :
+
+```ts
+OidcModule.forRootAsync({
+  useFactory: async (configService: ConfigService) => ({
+    issuer: configService.get('OIDC_ISSUER'),
+    clientMetadata: {
+      client_id: configService.get('OIDC_CLIENT_ID'),
+      client_secret: configService.get('OIDC_CLIENT_SECRET'),
+    },
+    authParams: {
+      scopes: configService.get('OIDC_SCOPES'),
+    },
+    origin: configService.get('ORIGIN'),
+  }),
+  inject: [ConfigService],
+  imports: [ConfigModule],
+});
+```
+
+If using a class, `createOidcOptions` was changed to a more generic `createModuleConfig`.
 
 ### Documentation